circular dependency issue fixed
