+++ template/src/routes/app-redirect.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
-
-import { goToPage, Params } from './route-helpers';
-import type { RouteName, RouteParams } from './routes';
+import { goToPage } from './route-helpers';
+import { Params, RouteName } from './utils';
+import type { RouteParams } from './routes';
 
 /*
   This component is a wrapper for redirecting across and inside apps.
