+++ template/src/routes/index.ts
@@ -1,5 +1,7 @@
+export { RouteComponent } from './route-component';
 export { AppLink } from './app-link';
 export { AppRedirect } from './app-redirect';
 export { goToPage } from './route-helpers';
-export { routes, RouteName } from './routes';
+export { routes } from './routes';
+export { RouteName } from './utils';
 export { Router } from './router';
