+++ template/src/routes/route-helpers.ts
@@ -1,10 +1,6 @@
 import { ParamsHelper } from 'helpers/params-helper';
-import { history } from './router';
-import { RouteName, routes } from './routes';
-
-export type Params = {
-  [key: string]: string | boolean | number | null,
-};
+import { routes } from './routes';
+import { Params, RouteName, history } from './utils';
 
 /*
   Given a name, finds the route in the configuration that has that name.
