+++ template/src/routes/routes.ts
@@ -1,34 +1,21 @@
 /* eslint-disable @typescript-eslint/no-shadow */
-import { Home } from 'pages/home';
-import { About } from 'pages/about';
-import { NotFound } from 'pages/not-found';
 import type { ExtractRouteParams } from 'react-router';
-import type { Params } from './route-helpers';
-import { setPathParams } from './utils';
-
-export enum RouteName {
-  Home = 'home',
-  About = 'about',
-  NotFound = 'notFound',
-}
+import { Params, RouteName, setPathParams } from './utils';
 
 const ROUTES = [
   {
     name: RouteName.Home,
-    path: '/:page?',
+    path: '/',
     exact: true,
-    component: Home,
   },
   {
     name: RouteName.About,
     path: '/about',
     exact: true,
-    component: About,
   },
   {
     name: RouteName.NotFound,
     path: '*',
-    component: NotFound,
   },
 ] as const;
